data/pipelines/requirements.txt
requests==2.32.3
httpx==0.27.2
pandas==2.2.3
geopandas==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
netCDF4==1.7.1
rasterio==1.3.11
boto3==1.35.23
redis==5.0.8
python-dotenv==1.0.1
